test: test uncoupled task with gym
